[FEATURE] 노트, 폴더 루트 변경 기능 추가

See merge request s09-final/S09P31A407!233

+++ frontend/app/components/userPage/SideBar/Directory/Folder.tsx
@@ -10,6 +10,7 @@ import { createFolder, deleteFolder } from '@api/folder/folder';
 import IconButton from '@components/common/IconButton';
 import SideDataContext from '@(pages)/user-page/[userNickname]/SideDataContext';
 import { showAlert } from '@src/util/alert';
+import Swal from 'sweetalert2';
 
 export interface FolderProps {
   type?: 'folder';
@@ -30,6 +31,7 @@ export interface FolderProps {
   folderName?: string;
   setFolderName?: Dispatch<SetStateAction<string>>;
   isModifyDelete?: boolean;
+  onNoteEdit: (editingNoteId: number, title: string) => void;
 }
 
 export default function Folder({
@@ -47,6 +49,7 @@ export default function Folder({
   setFolderName,
   isModifyDelete,
   onEdit,
+  onNoteEdit,
 }: FolderProps) {
   const paddingLeft = `${depth * 1.25}rem`;
   const [isFolderOpen, setFolderOpen] = useState(
@@ -98,17 +101,39 @@ export default function Folder({
   };
 
   // 폴더 삭제
-  const handleDelete = async () => {
-    if (getSideList) {
-      try {
-        await deleteFolder(id);
-        getSideList();
-        getGraphData();
-        getNoteGraphData();
-      } catch {
-        showAlert('폴더 삭제에 실패했습니다', 'error');
+  const handleDelete = async (title: string) => {
+    const textColor = theme === 'light' ? 'black' : 'white';
+    Swal.fire({
+      html: `${title} 폴더를<br/>삭제하시겠습니까?`,
+      showCloseButton: true,
+      width: 300,
+      background:
+        theme === 'light' ? colors.modal : colors['dark-background-page'],
+      color: textColor,
+      confirmButtonText: '삭제하기',
+      confirmButtonColor: colors.warn,
+      showClass: {
+        backdrop: 'swal2-noanimation',
+        popup: '',
+        icon: '',
+      },
+      hideClass: {
+        popup: '',
+      },
+    }).then(async (res) => {
+      if (res.isConfirmed) {
+        if (getSideList) {
+          try {
+            await deleteFolder(id);
+            getSideList();
+            getGraphData();
+            getNoteGraphData();
+          } catch {
+            showAlert('폴더 삭제에 실패했습니다', 'error');
+          }
+        }
       }
-    }
+    });
   };
 
   // 폴더아이디 상위로 전달
@@ -151,7 +176,7 @@ export default function Folder({
             />
             <IconButton
               size={18}
-              onClick={handleDelete}
+              onClick={() => handleDelete(title)}
               theme={theme}
               name="Remove"
             />
@@ -175,6 +200,7 @@ export default function Folder({
                   setParentId={setParentId}
                   currentNoteId={currentNoteId}
                   isModifyDelete={isModifyDelete}
+                  onNoteEdit={onNoteEdit}
                 />
               ) : (
                 <Folder
@@ -194,6 +220,7 @@ export default function Folder({
                   setFolderName={setFolderName}
                   isModifyDelete={isModifyDelete}
                   onEdit={onEdit}
+                  onNoteEdit={onNoteEdit}
                 />
               )
             )}
@@ -206,6 +233,7 @@ export default function Folder({
           <div style={{ paddingLeft: '1.25rem' }}>
             <CreateFile
               theme={theme}
+              placeholder="폴더이름"
               onChange={(e) => setFolderName && setFolderName(e.target.value)}
               onKeyDown={(e) => handleKeyDown && handleKeyDown(e)}
               type="folder"
